Fix various features that didn't work if /usr wasn't mounted in the initramfs.
Amongst other things, this commit drops the special-casing of /etc/localtime
when /usr is not part of the root fs, as timedated does not care about the
file content, only the symlink value.

To fix the real problem, that timezoned needs zoneinfo data which is located
in /usr, add RequiresMountsFor=usr/share/zoneinfo to the service file.