Add Pre-Depends on systemd to systemd-sysv.
This avoid risking that the sysv-compatible symlinks become dangling on a partial install.
Note that the Pre-Depends is unversioned to facilitate upgrades to newer systemd versions,
and that the versioned Depends is kept to ensure that user doesn't get confused to thinking
they are running a different systemd version then they are.